ZZ: source-tree: sync with blu_spark branch https://github.com/engstk/op6/commits/blu_spark-10
Obviously some files in techpack/ got deleted.I couldn't find a reason (there is no commit!?)
Maybe it's for fixing the 'missing Makefile' build errors which occur if they are kept in place?
The miracle remains how the changes went in w/o a commit but anyways removing done here too to be in
sync with original source.